% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dtc_import}
\alias{dtc_import}
\title{Bulk import traffic data}
\usage{
dtc_import(dir = tempdir())
}
\arguments{
\item{dir}{Where should the data be stored? `tempdir()` by default.}
}
\description{
This function downloads and reads bulk traffic data from the DfT
}
